Add reconnection thread and tweak params
Using another thread to reconnect to peers that we disconnect so we can
assess their lifetime. In this stage, we use a queue of addresses we
want to reconnect periodically. For now, all disconnected nodes will be
added to the queue. (Probably have to change that behavior later)

Network parameters are changed to make the scans a little bit more
"active". By doing so, we have to allow more outgoing connections and
total connections.